Combine parsing of mirror and scanner results
