+++ src/main/java/org/dependencytrack/event/kafka/processor/MirrorVulnerabilityProcessor.java
@@ -10,11 +10,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.kafka.streams.processor.api.Processor;
 import org.apache.kafka.streams.processor.api.Record;
 import org.cyclonedx.proto.v1_4.Bom;
-import org.cyclonedx.proto.v1_4.ScoreMethod;
-import org.cyclonedx.proto.v1_4.Source;
 import org.dependencytrack.event.IndexEvent;
-import org.dependencytrack.model.Cwe;
-import org.dependencytrack.model.Severity;
+import org.dependencytrack.parser.hyades.ModelConverterCdxToVuln;
 import org.dependencytrack.model.Vulnerability;
 import org.dependencytrack.model.VulnerableSoftware;
 import org.dependencytrack.parser.nvd.ModelConverter;
@@ -22,9 +19,6 @@ import org.dependencytrack.persistence.QueryManager;
 import us.springett.parsers.cpe.exceptions.CpeEncodingException;
 import us.springett.parsers.cpe.exceptions.CpeParsingException;
 
-import java.math.BigDecimal;
-import java.sql.Date;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -48,7 +42,7 @@ public class MirrorVulnerabilityProcessor implements Processor<String, Bom, Void
             String key = record.key();
             String mirrorSource = key.substring(0, key.indexOf("/"));
             Vulnerability.Source source = Vulnerability.Source.valueOf(mirrorSource);
-            final Vulnerability vulnerability = mapBomToVulnerability(qm, bom);
+            final Vulnerability vulnerability = ModelConverterCdxToVuln.convert(qm, bom, bom.getVulnerabilities(0), false);
             final List<VulnerableSoftware> vsListOld = qm.detach(qm.getVulnerableSoftwareByVulnId(vulnerability.getSource(), vulnerability.getVulnId()));
             final Vulnerability synchronizedVulnerability = qm.synchronizeVulnerability(vulnerability, false);
             var cycloneVuln = bom.getVulnerabilities(0);
@@ -259,87 +253,5 @@ public class MirrorVulnerabilityProcessor implements Processor<String, Bom, Void
         vs.setVersionEndExcluding(versionEndExcluding);
         vs.setVersionEndIncluding(versionEndIncluding);
         return vs;
-    }
-
-    public Vulnerability mapBomToVulnerability(final QueryManager qm, final Bom bom) {
-        org.cyclonedx.proto.v1_4.Vulnerability cycloneVuln = bom.getVulnerabilities(0);
-        final Vulnerability vuln = new Vulnerability();
-        if (cycloneVuln.getId() != null) {
-            vuln.setSource(extractSource(cycloneVuln.getId(), cycloneVuln.getSource()));
-        }
-        vuln.setVulnId(cycloneVuln.getId());
-        vuln.setTitle(StringUtils.abbreviate(cycloneVuln.getDescription(), 255));
-        vuln.setDescription(cycloneVuln.getDetail());
-        vuln.setPublished(Date.from(Instant.ofEpochSecond(cycloneVuln.getPublished().getSeconds())));
-        vuln.setUpdated(Date.from(Instant.ofEpochSecond(cycloneVuln.getUpdated().getSeconds())));
-
-        if (cycloneVuln.getCredits() != null) {
-            vuln.setCredits(String.join(", ", cycloneVuln.getCredits().toString()));
-        }
-        final StringBuilder sb = new StringBuilder();
-        if (!bom.getExternalReferencesList().isEmpty()) {
-            bom.getExternalReferencesList().stream().forEach(externalReference -> {
-                sb.append("* [").append(externalReference.getUrl()).append("](").append(externalReference.getUrl()).append(")\n");
-            });
-            vuln.setReferences(sb.toString());
-        }
-        if(!cycloneVuln.getAdvisoriesList().isEmpty()){
-            cycloneVuln.getAdvisoriesList().stream().forEach(advisory -> {
-                sb.append("* [").append(advisory.getUrl()).append("](").append(advisory.getUrl()).append(")\n");
-            });
-            vuln.setReferences(sb.toString());
-        }
-
-
-        if (!cycloneVuln.getCwesList().isEmpty()) {
-            cycloneVuln.getCwesList().stream().forEach(cweId -> {
-                final Cwe cwe = qm.getCweById(cweId);
-                if (cwe != null) {
-                    vuln.addCwe(cwe);
-                }
-            });
-        }
-        vuln.setSeverity(calculateSeverity(bom));
-        if (!cycloneVuln.getRatingsList().isEmpty() && cycloneVuln.getRatings(0).getMethod() != null) {
-            if (cycloneVuln.getRatings(0).getMethod().equals(ScoreMethod.SCORE_METHOD_CVSSV2)) {
-                vuln.setCvssV2Vector(cycloneVuln.getRatings(0).getVector());
-                vuln.setCvssV2BaseScore(BigDecimal.valueOf(cycloneVuln.getRatings(0).getScore()));
-            }
-            if (cycloneVuln.getRatings(0).getMethod().equals(ScoreMethod.SCORE_METHOD_CVSSV3)) {
-                vuln.setCvssV3Vector(cycloneVuln.getRatings(0).getVector());
-                vuln.setCvssV3BaseScore(BigDecimal.valueOf(cycloneVuln.getRatings(0).getScore()));
-            }
-        }
-        return vuln;
-    }
-
-    public Severity calculateSeverity(Bom bom) {
-        if (bom.getVulnerabilities(0) != null
-                && !bom.getVulnerabilities(0).getRatingsList().isEmpty()
-                && bom.getVulnerabilities(0).getRatings(0) != null) {
-            org.cyclonedx.proto.v1_4.Severity severity =
-                    bom.getVulnerabilities(0).getRatings(0).getSeverity();
-            if (severity != null) {
-                if (severity.equals(org.cyclonedx.proto.v1_4.Severity.SEVERITY_CRITICAL)) {
-                    return Severity.CRITICAL;
-                } else if (severity.equals(org.cyclonedx.proto.v1_4.Severity.SEVERITY_HIGH)) {
-                    return Severity.HIGH;
-                } else if (severity.equals(org.cyclonedx.proto.v1_4.Severity.SEVERITY_MEDIUM)) {
-                    return Severity.MEDIUM;
-                } else if (severity.equals(org.cyclonedx.proto.v1_4.Severity.SEVERITY_LOW)) {
-                    return Severity.LOW;
-                }
-            }
-        }
-        return Severity.UNASSIGNED;
-    }
-
-    public Vulnerability.Source extractSource(String vulnId, Source source) {
-        final String sourceId = vulnId.split("-")[0];
-        return switch (sourceId) {
-            case "GHSA" -> Vulnerability.Source.GITHUB;
-            case "CVE" -> Vulnerability.Source.NVD;
-            default -> source!= null ? Vulnerability.Source.valueOf(source.getName()) : Vulnerability.Source.INTERNAL;
-        };
     }
 }
