+++ src/main/java/org/dependencytrack/event/kafka/processor/VulnerabilityScanResultProcessor.java
@@ -23,6 +23,7 @@ import org.dependencytrack.notification.NotificationGroup;
 import org.dependencytrack.notification.NotificationScope;
 import org.dependencytrack.notification.vo.NewVulnerabilityIdentified;
 import org.dependencytrack.notification.vo.NewVulnerableDependency;
+import org.dependencytrack.parser.hyades.ModelConverterCdxToVuln;
 import org.dependencytrack.persistence.QueryManager;
 import org.dependencytrack.util.PersistenceUtil;
 import org.hyades.proto.vulnanalysis.v1.ScanKey;
@@ -32,7 +33,6 @@ import org.hyades.proto.vulnanalysis.v1.Scanner;
 import org.hyades.proto.vulnanalysis.v1.ScannerResult;
 
 import javax.jdo.Query;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
@@ -40,7 +40,7 @@ import java.util.NoSuchElementException;
 import java.util.Set;
 import java.util.UUID;
 
-import static org.dependencytrack.parser.hyades.ModelConverter.convert;
+import static org.dependencytrack.parser.hyades.ModelConverterCdxToVuln.convert;
 import static org.dependencytrack.util.NotificationUtil.generateNotificationContent;
 import static org.dependencytrack.util.NotificationUtil.generateNotificationTitle;
 import static org.dependencytrack.util.PersistenceUtil.applyIfChanged;
@@ -160,10 +160,10 @@ public class VulnerabilityScanResultProcessor extends ContextualFixedKeyProcesso
     private Set<Vulnerability> syncVulnerabilities(final QueryManager qm, final ScanKey scanKey, final ScannerResult scannerResult) {
         final var syncedVulns = new HashSet<Vulnerability>();
 
-        for (final org.hyades.proto.vuln.v1.Vulnerability reportedVuln : scannerResult.getVulnerabilitiesList()) {
+        for (final org.cyclonedx.proto.v1_4.Vulnerability reportedVuln : scannerResult.getBom().getVulnerabilitiesList()) {
             final Vulnerability vuln;
             try {
-                vuln = convert(reportedVuln);
+                vuln = ModelConverterCdxToVuln.convert(qm, scannerResult.getBom(), reportedVuln, true);
             } catch (RuntimeException e) {
                 LOGGER.error("Failed to convert vulnerability %s/%s (reported by %s for component %s) to internal model (scanKey: %s)"
                         .formatted(reportedVuln.getSource(), reportedVuln.getId(), scannerResult.getScanner(), scanKey.getComponentUuid(), prettyPrint(scanKey)), e);
