--- src/main/java/org/dependencytrack/parser/ossindex/model/ComponentReportVulnerability.java
@@ -1,113 +1,0 @@
-/*
- * This file is part of Dependency-Track.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * Copyright (c) Steve Springett. All Rights Reserved.
- */
-package org.dependencytrack.parser.ossindex.model;
-
-import java.util.List;
-
-/**
- * The response from Sonatype OSS Index will respond with 0 or more ComponentReports. This
- * class defines the ComponentReportVulnerability objects returned within each ComponentReport.
- *
- * @author Steve Springett
- * @since 3.2.0
- */
-public class ComponentReportVulnerability {
-
-    private String id;
-    private String title;
-    private String description;
-    private Number cvssScore;
-    private String cvssVector;
-    private String cwe;
-    private String cve;
-    private String reference;
-    private List<String> externalReferences;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(final String id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(final String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(final String description) {
-        this.description = description;
-    }
-
-    public Number getCvssScore() {
-        return cvssScore;
-    }
-
-    public void setCvssScore(final Number cvssScore) {
-        this.cvssScore = cvssScore;
-    }
-
-    public String getCvssVector() {
-        return cvssVector;
-    }
-
-    public void setCvssVector(final String cvssVector) {
-        this.cvssVector = cvssVector;
-    }
-
-    public String getCwe() {
-        return cwe;
-    }
-
-    public void setCwe(final String cwe) {
-        this.cwe = cwe;
-    }
-
-    public String getCve() {
-        return cve;
-    }
-
-    public void setCve(final String cve) {
-        this.cve = cve;
-    }
-
-    public String getReference() {
-        return reference;
-    }
-
-    public void setReference(final String reference) {
-        this.reference = reference;
-    }
-
-    public List<String> getExternalReferences() {
-        return externalReferences;
-    }
-
-    public void setExternalReferences(final List<String> externalReferences) {
-        this.externalReferences = externalReferences;
-    }
-}
