+++ src/test/java/org/dependencytrack/event/kafka/processor/MirrorVulnerabilityProcessorTest.java
@@ -6,6 +6,7 @@ import org.dependencytrack.PersistenceCapableTest;
 import org.dependencytrack.model.Severity;
 import org.dependencytrack.model.Vulnerability;
 import org.dependencytrack.model.VulnerableSoftware;
+import org.dependencytrack.parser.hyades.ModelConverterCdxToVuln;
 import org.dependencytrack.util.KafkaTestUtil;
 import org.junit.Assert;
 import org.junit.Before;
@@ -99,16 +100,16 @@ public class MirrorVulnerabilityProcessorTest extends PersistenceCapableTest {
 
     @Test
     public void testVulnRangeParsing(){
-        MirrorVulnerabilityProcessor processor = new MirrorVulnerabilityProcessor();
-        Vulnerability vuln = processor.mapBomToVulnerability(qm, bom);
+        Vulnerability vuln = ModelConverterCdxToVuln.convert(qm, bom, bom.getVulnerabilities(0), false);
         Assert.assertNotNull(vuln);
         Assert.assertEquals("GHSA-xmw9-q7x9-j5qc",vuln.getVulnId());
         Assert.assertEquals("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H", vuln.getCvssV3Vector());
-        Severity severity = processor.calculateSeverity(bom);
+        Severity severity = ModelConverterCdxToVuln.calculateSeverity(bom);
         Assert.assertEquals("HIGH", severity.name());
-        Vulnerability.Source source = processor.extractSource("GHSA-xmw9-q7x9-j5qc", bom.getVulnerabilities(0).getSource());
+        Vulnerability.Source source = ModelConverterCdxToVuln.extractSource("GHSA-xmw9-q7x9-j5qc", bom.getVulnerabilities(0).getSource());
         Assert.assertEquals("GITHUB", source.name());
 
+        MirrorVulnerabilityProcessor processor = new MirrorVulnerabilityProcessor();
         VulnerableSoftware vs = processor.mapAffectedRangeToVulnerableSoftware(qm, bom.getVulnerabilities(0).getAffects(0).getVersions(0).getRange(), "pkg:maven/org.http4s/blaze-core_2.11", null);
         Assert.assertNotNull(vs);
         Assert.assertEquals("0", vs.getVersionStartIncluding());
@@ -126,16 +127,16 @@ public class MirrorVulnerabilityProcessorTest extends PersistenceCapableTest {
 
     @Test
     public void testVulnVersionParsing(){
-        MirrorVulnerabilityProcessor processor = new MirrorVulnerabilityProcessor();
-        Vulnerability vuln = processor.mapBomToVulnerability(qm, bom);
+        Vulnerability vuln = ModelConverterCdxToVuln.convert(qm, bom, bom.getVulnerabilities(0), false);
         Assert.assertNotNull(vuln);
         Assert.assertEquals("GHSA-xmw9-q7x9-j5qc",vuln.getVulnId());
         Assert.assertEquals("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H", vuln.getCvssV3Vector());
-        Severity severity = processor.calculateSeverity(bom);
+        Severity severity = ModelConverterCdxToVuln.calculateSeverity(bom);
         Assert.assertEquals("HIGH", severity.name());
-        Vulnerability.Source source = processor.extractSource("GHSA-xmw9-q7x9-j5qc", bom.getVulnerabilities(0).getSource());
+        Vulnerability.Source source = ModelConverterCdxToVuln.extractSource("GHSA-xmw9-q7x9-j5qc", bom.getVulnerabilities(0).getSource());
         Assert.assertEquals("GITHUB", source.name());
 
+        MirrorVulnerabilityProcessor processor = new MirrorVulnerabilityProcessor();
         var vs = processor.mapAffectedVersionToVulnerableSoftware(qm, bom.getVulnerabilities(0).getAffects(0).getVersions(1).getVersion(), "pkg:maven/org.http4s/blaze-core_2.11", null);
         Assert.assertNotNull(vs);
         Assert.assertEquals("0.10.0", vs.getVersion());
@@ -150,9 +151,8 @@ public class MirrorVulnerabilityProcessorTest extends PersistenceCapableTest {
 
     @Test
     public void testExtractSource() {
-        MirrorVulnerabilityProcessor processor = new MirrorVulnerabilityProcessor();
         Source source = Source.newBuilder().setName("OSV").build();
-        Assert.assertEquals(Vulnerability.Source.OSV, processor.extractSource("OSV-test-id", source));
+        Assert.assertEquals(Vulnerability.Source.OSV, ModelConverterCdxToVuln.extractSource("OSV-test-id", source));
     }
 
 }
