+++ src/test/java/org/dependencytrack/event/kafka/processor/VulnerabilityScanResultProcessorTest.java
@@ -12,6 +12,10 @@ import org.apache.kafka.streams.TopologyTestDriver;
 import org.apache.kafka.streams.kstream.Consumed;
 import org.apache.kafka.streams.kstream.Produced;
 import org.apache.kafka.streams.test.TestRecord;
+import org.cyclonedx.proto.v1_4.Advisory;
+import org.cyclonedx.proto.v1_4.Bom;
+import org.cyclonedx.proto.v1_4.Source;
+import org.cyclonedx.proto.v1_4.VulnerabilityRating;
 import org.dependencytrack.PersistenceCapableTest;
 import org.dependencytrack.event.kafka.KafkaEventHeaders;
 import org.dependencytrack.event.kafka.KafkaTopics;
@@ -31,10 +35,6 @@ import org.dependencytrack.persistence.CweImporter;
 import org.hyades.proto.notification.v1.NewVulnerabilitySubject;
 import org.hyades.proto.notification.v1.NewVulnerableDependencySubject;
 import org.hyades.proto.notification.v1.Notification;
-import org.hyades.proto.vuln.v1.Rating;
-import org.hyades.proto.vuln.v1.Reference;
-import org.hyades.proto.vuln.v1.ScoreMethod;
-import org.hyades.proto.vuln.v1.Source;
 import org.hyades.proto.vulnanalysis.v1.ScanKey;
 import org.hyades.proto.vulnanalysis.v1.ScanResult;
 import org.hyades.proto.vulnanalysis.v1.Scanner;
@@ -51,6 +51,9 @@ import java.util.List;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.cyclonedx.proto.v1_4.ScoreMethod.SCORE_METHOD_CVSSV2;
+import static org.cyclonedx.proto.v1_4.ScoreMethod.SCORE_METHOD_CVSSV3;
+import static org.cyclonedx.proto.v1_4.ScoreMethod.SCORE_METHOD_OWASP;
 import static org.dependencytrack.util.KafkaTestUtil.deserializeValue;
 import static org.hyades.proto.notification.v1.Group.GROUP_ANALYZER;
 import static org.hyades.proto.notification.v1.Group.GROUP_NEW_VULNERABILITY;
@@ -59,9 +62,6 @@ import static org.hyades.proto.notification.v1.Level.LEVEL_ERROR;
 import static org.hyades.proto.notification.v1.Level.LEVEL_INFORMATIONAL;
 import static org.hyades.proto.notification.v1.Scope.SCOPE_PORTFOLIO;
 import static org.hyades.proto.notification.v1.Scope.SCOPE_SYSTEM;
-import static org.hyades.proto.vuln.v1.Source.SOURCE_INTERNAL;
-import static org.hyades.proto.vuln.v1.Source.SOURCE_NVD;
-import static org.hyades.proto.vuln.v1.Source.SOURCE_OSSINDEX;
 import static org.hyades.proto.vulnanalysis.v1.ScanStatus.SCAN_STATUS_FAILED;
 import static org.hyades.proto.vulnanalysis.v1.ScanStatus.SCAN_STATUS_PENDING;
 import static org.hyades.proto.vulnanalysis.v1.ScanStatus.SCAN_STATUS_SUCCESSFUL;
@@ -186,9 +186,7 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                 .addScannerResults(ScannerResult.newBuilder()
                         .setScanner(SCANNER_INTERNAL)
                         .setStatus(SCAN_STATUS_SUCCESSFUL)
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId("INT-001")
-                                .setSource(SOURCE_INTERNAL)))
+                        .setBom(Bom.newBuilder().addVulnerabilities(createVuln("INT-001", "INTERNAL"))))
                 .build();
 
         inputTopic.pipeInput(scanKey, scanResult);
@@ -229,15 +227,11 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                 .addScannerResults(ScannerResult.newBuilder()
                         .setScanner(SCANNER_INTERNAL)
                         .setStatus(SCAN_STATUS_SUCCESSFUL)
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId(vulnA.getVulnId())
-                                .setSource(SOURCE_INTERNAL))
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId(vulnB.getVulnId())
-                                .setSource(SOURCE_OSSINDEX))
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId(vulnC.getVulnId())
-                                .setSource(SOURCE_INTERNAL)))
+                        .setBom(Bom.newBuilder().addAllVulnerabilities(List.of(
+                                createVuln(vulnA.getVulnId(), "INTERNAL"),
+                                createVuln(vulnB.getVulnId(), "OSSINDEX"),
+                                createVuln(vulnC.getVulnId(), "INTERNAL")
+                                ))))
                 .build();
         final Headers headers = new RecordHeaders();
         headers.add(KafkaEventHeaders.VULN_ANALYSIS_LEVEL, VulnerabilityAnalysisLevel.BOM_UPLOAD_ANALYSIS.name().getBytes());
@@ -338,9 +332,7 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                 .addScannerResults(ScannerResult.newBuilder()
                         .setScanner(SCANNER_INTERNAL)
                         .setStatus(SCAN_STATUS_SUCCESSFUL)
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId("CVE-001")
-                                .setSource(SOURCE_NVD)))
+                        .setBom(Bom.newBuilder().addVulnerabilities(createVuln("CVE-001", "NVD"))))
                 .build();
 
         inputTopic.pipeInput(scanKey, scanResult);
@@ -438,10 +430,9 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                 .addScannerResults(ScannerResult.newBuilder()
                         .setScanner(scanner)
                         .setStatus(SCAN_STATUS_SUCCESSFUL)
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                .setId(vulnId)
-                                .setSource(Source.valueOf("SOURCE_" + vulnSource))
-                                .setDescription("modified description")))
+                        .setBom(Bom.newBuilder().addVulnerabilities(
+                                createVuln(vulnId, vulnSource).toBuilder().setDetail("modified description")
+                                        .build())))
                 .build();
 
         inputTopic.pipeInput(scanKey, scanResult);
@@ -510,15 +501,11 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                 .addScannerResults(ScannerResult.newBuilder()
                         .setScanner(SCANNER_OSSINDEX)
                         .setStatus(SCAN_STATUS_SUCCESSFUL)
-                        .addVulnerabilities(org.hyades.proto.vuln.v1.Vulnerability.newBuilder()
-                                        .setId("CVE-001")
-                                        .setSource(SOURCE_NVD)
-                                        .setTitle("modified title")
-                                        // (There is no subTitle field)
-                                        .setDescription("modified description")
+                        .setBom(Bom.newBuilder().addVulnerabilities(
+                                createVuln("CVE-001", "NVD").toBuilder()
+                                        .setDescription("modified title")
+                                        .setDetail("modified description")
                                         .setRecommendation("modified recommendation")
-                                        .addReferences(Reference.newBuilder()
-                                                .setUrl("modified reference"))
                                         .setCreated(Timestamp.newBuilder()
                                                 .setSeconds(1673305200)) // Mon Jan 09 2023 23:00:00 GMT+0000
                                         .setPublished(Timestamp.newBuilder()
@@ -526,26 +513,23 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
                                         .setUpdated(Timestamp.newBuilder()
                                                 .setSeconds(1673478000)) // Wed Jan 11 2023 23:00:00 GMT+0000
                                         .addCwes(999)
-                                        .addRatings(Rating.newBuilder()
-                                                .setSource(SOURCE_NVD)
-                                                .setMethod(ScoreMethod.SCORE_METHOD_CVSSV2)
+                                        .addAdvisories(Advisory.newBuilder().setUrl("modified reference").build())
+                                        .addRatings(VulnerabilityRating.newBuilder()
+                                                .setSource(Source.newBuilder().setName("NVD").build())
+                                                .setMethod(SCORE_METHOD_CVSSV2)
                                                 .setScore(9.3)
                                                 .setVector("(AV:N/AC:M/Au:N/C:C/I:C/A:C)"))
-                                        .addRatings(Rating.newBuilder()
-                                                .setSource(SOURCE_NVD)
-                                                .setMethod(ScoreMethod.SCORE_METHOD_CVSSV31)
+                                        .addRatings(VulnerabilityRating.newBuilder()
+                                                .setSource(Source.newBuilder().setName("NVD").build())
+                                                .setMethod(SCORE_METHOD_CVSSV3)
                                                 .setScore(10.0)
-                                                .setVector("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:C/C:H/I:H/A:H"))
-                                        .addRatings(Rating.newBuilder()
-                                                .setSource(SOURCE_OSSINDEX)
-                                                .setMethod(ScoreMethod.SCORE_METHOD_OWASP)
+                                                .setVector("CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:C/C:H/I:H/A:H"))
+                                        .addRatings(VulnerabilityRating.newBuilder()
+                                                .setSource(Source.newBuilder().setName("OSSINDEX").build())
+                                                .setMethod(SCORE_METHOD_OWASP)
                                                 .setScore(6.6)
                                                 .setVector("SL:1/M:4/O:4/S:9/ED:7/EE:3/A:4/ID:3/LC:9/LI:1/LAV:5/LAC:1/FD:3/RD:4/NC:7/PV:9"))
-                                // (There is no vulnerableVersions field)
-                                // (There is no patchedVersions field)
-                                // (There is no epssScore field)
-                                // (There is no epssPercentile field)
-                        ))
+                                        .build())))
                 .build();
 
         inputTopic.pipeInput(scanKey, scanResult);
@@ -559,7 +543,7 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
         assertThat(vulnerability.getTitle()).isEqualTo("modified title");
         assertThat(vulnerability.getSubTitle()).isNull();
         assertThat(vulnerability.getRecommendation()).isEqualTo("modified recommendation");
-        assertThat(vulnerability.getReferences()).isEqualTo("* [modified reference](modified reference)");
+        assertThat(vulnerability.getReferences()).isEqualTo("* [modified reference](modified reference)\n");
         assertThat(vulnerability.getCreated()).isEqualTo(Instant.ofEpochSecond(1673305200));
         assertThat(vulnerability.getPublished()).isEqualTo(Instant.ofEpochSecond(1673391600));
         assertThat(vulnerability.getUpdated()).isEqualTo(Instant.ofEpochSecond(1673478000));
@@ -580,5 +564,11 @@ public class VulnerabilityScanResultProcessorTest extends PersistenceCapableTest
         assertThat(vulnerability.getPatchedVersions()).isNull();
         assertThat(vulnerability.getEpssScore()).isEqualByComparingTo("5.1");
         assertThat(vulnerability.getEpssPercentile()).isEqualByComparingTo("5.2");
+    }
+
+    private org.cyclonedx.proto.v1_4.Vulnerability createVuln(final String id, final String source) {
+        return org.cyclonedx.proto.v1_4.Vulnerability.newBuilder()
+                .setId(id)
+                .setSource(Source.newBuilder().setName(source).build()).build();
     }
 }
